// src/api/generated/schemas/getRoles200DataItem.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Elysia Documentation
 * Development documentation
 * OpenAPI spec version: 0.0.0
 */
import type { GetRoles200DataItemDescription } from './getRoles200DataItemDescription';
import type { GetRoles200DataItemPortal } from './getRoles200DataItemPortal';
import type { GetRoles200DataItemCreatedAt } from './getRoles200DataItemCreatedAt';
import type { GetRoles200DataItemUpdatedAt } from './getRoles200DataItemUpdatedAt';

export type GetRoles200DataItem = {
  id: string;
  slug: string;
  name: string;
  /** @nullable */
  description: GetRoles200DataItemDescription;
  portal: GetRoles200DataItemPortal;
  isSystem: boolean;
  createdAt: GetRoles200DataItemCreatedAt;
  updatedAt: GetRoles200DataItemUpdatedAt;
  userCount: number;
  permissionIds: string[];
};